+++ api/src/main/java/org/cyanteam/telemaniacs/core/dto/TransmissioinDTO.java
@@ -1,0 +1,7 @@
+package org.cyanteam.telemaniacs.core.dto;
+
+/**
+ * Created by tinko on 11/25/2017.
+ */
+public class TransmissioinDTO {
+}
