+++ api/src/main/java/org/cyanteam/telemaniacs/core/facade/ScheduleFacade.java
@@ -1,0 +1,7 @@
+package org.cyanteam.telemaniacs.core.facade;
+
+/**
+ * Created by tinko on 11/25/2017.
+ */
+public interface ScheduleFacade {
+}
