+++ persistence/src/main/java/org/cyanteam/telemaniacs/core/dto/TransmissionDao.java
@@ -1,6 +1,5 @@
-package org.cyanteam.telemaniacs.core.dao;
+package org.cyanteam.telemaniacs.core.dto;
 
-import org.cyanteam.telemaniacs.core.entities.Channel;
 import org.cyanteam.telemaniacs.core.entities.Transmission;
 
 import javax.validation.ConstraintViolationException;
