+++ persistence/src/main/java/org/cyanteam/telemaniacs/core/dto/TransmissionOccurrenceDaoImpl.java
@@ -1,4 +1,4 @@
-package org.cyanteam.telemaniacs.core.dao;
+package org.cyanteam.telemaniacs.core.dto;
 
 import org.cyanteam.telemaniacs.core.entities.Channel;
 import org.cyanteam.telemaniacs.core.entities.Transmission;
