+++ persistence/src/main/java/org/cyanteam/telemaniacs/core/dto/VotingDao.java
@@ -1,4 +1,4 @@
-package org.cyanteam.telemaniacs.core.dao;
+package org.cyanteam.telemaniacs.core.dto;
 
 import java.util.List;
 import javax.validation.ConstraintViolationException;
